reworked 7 task

+++ src/week1/hometask1/_07HomeTask_1.java
@@ -11,13 +11,13 @@ public class _07HomeTask_1 {
     public static void main(String[] args) {
         Scanner scanout = new Scanner(System.in);
         System.out.println("Enter first number...");
-        double number1 = scanout.nextDouble();
+        int number1 = scanout.nextInt();
         System.out.println("Enter second number...");
-        double number2 = scanout.nextDouble();
+        int number2 = scanout.nextInt();
 
-        double result = number1 / number2;
-        int resultPart1 = (int) result;
-        double resultPart2 = number1 % number2;
+        int result = number1 / number2;
+        int resultPart1 = result;
+        int resultPart2 = number1 % number2;
 
         if (resultPart2 == 0){
             System.out.println("True " + resultPart1 + " and " + resultPart2);
